new_project/apps/bot_gateway/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rpi-trader-bot-gateway"
version = "0.1.0"
description = "Telegram Bot Gateway for RPI Trader"
dependencies = [
    "python-telegram-bot>=20.7",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "psutil>=5.9.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
]

[project.scripts]
bot-gateway = "bot_gateway.main:main"



[tool.hatch.build.targets.wheel]
packages = ["bot_gateway"]

